integrations/vapi-voice/requirements.txt
vapi_server_sdk
python-dotenv
